+++ .gitignore
@@ -131,3 +131,5 @@ dmypy.json
 .idea/
 
 private-data/
+
+.DS_Store
