+++ nyc_data/ppe/aggregations.py
@@ -6,7 +6,7 @@ import django_tables2 as tables
 from django.utils.html import format_html
 
 import ppe.dataclasses as dc
-from ppe.models import Delivery, Inventory, ImportStatus
+from ppe.models import ScheduledDelivery, Inventory, ImportStatus
 
 
 @dataclass
@@ -40,7 +40,7 @@ def asset_rollup(
         rollup_fn: Callable[[dc.Item], any] = lambda x: x,
         estimate_demand=True
 ) -> Dict[str, AssetRollup]:
-    relevant_deliveries = Delivery.objects.prefetch_related("purchase", "source").filter(
+    relevant_deliveries = ScheduledDelivery.objects.prefetch_related("purchase", "source").filter(
         delivery_date__gte=time_start, delivery_date__lte=time_end, source__status=ImportStatus.active
     )
