+++ nyc_data/ppe/data_import.py
@@ -1,36 +1,35 @@
-import collections
-import dataclasses as dc
 import hashlib
 import tempfile
 from pathlib import Path
-from typing import Optional
+from typing import Optional, List
 
 import xlsx_utils
-from ppe.data_mappings import DataSource
-from ppe import data_mappings
-from ppe.models import Purchase, Delivery, Inventory, ImportStatus, DataImport
+from ppe.data_mapping.mappers import dcas_make, dcas_sourcing, inventory
+from ppe.data_mapping.types import DataFile
+from ppe.data_mapping.utils import ErrorCollector
+from ppe.models import ImportStatus, DataImport
 from xlsx_utils import import_xlsx
 
-MAPPINGS = {
-    DataSource.EDC_MAKE: data_mappings.SUPPLIERS_AND_PARTNERS,
-    DataSource.INVENTORY: data_mappings.INVENTORY,
-    DataSource.EDC_PPE: data_mappings.DCAS_DAILY_SOURCING
-}
+ALL_MAPPINGS = [
+    dcas_make.SUPPLIERS_AND_PARTNERS,
+    dcas_sourcing.DCAS_DAILY_SOURCING,
+    inventory.INVENTORY
+]
 
 
 class DataImportError(Exception):
     pass
 
 
-def handle_upload(f) -> DataImport:
+def handle_upload(f, uploader_name: str) -> DataImport:
     with tempfile.NamedTemporaryFile('w+b', delete=False, suffix=f.name) as upload_target:
         for chunk in f.chunks():
             upload_target.write(chunk)
-        return smart_import(Path(upload_target.name))
+        return smart_import(Path(upload_target.name), uploader_name)
 
 
-def import_in_progress(data_source: DataSource):
-    return DataImport.objects.filter(data_source=data_source, status=ImportStatus.candidate)
+def import_in_progress(data_file: DataFile):
+    return DataImport.objects.filter(data_file=data_file, status=ImportStatus.candidate)
 
 
 class NoMappingForFileError(DataImportError):
@@ -46,20 +45,21 @@ class ImportInProgressError(DataImportError):
         self.import_id = import_id
 
 
-def smart_import(path: Path) -> DataImport:
-    possible_mappings = xlsx_utils.guess_mapping(path, list(MAPPINGS.values()))
+def smart_import(path: Path, uploader_name: str, overwrite_in_prog: bool=False) -> DataImport:
+    possible_mappings = xlsx_utils.guess_mapping(path, ALL_MAPPINGS)
     if len(possible_mappings) == 0:
         raise NoMappingForFileError()
-    elif len(possible_mappings) > 1:
-        raise MultipleMappingsForFileError()
-    else:
-        inferred_mapping = possible_mappings[0]
-        data_source = [src for (src, mapping) in MAPPINGS.items() if mapping == inferred_mapping][0]
-        return import_data(path, data_source)
+    return import_data(path, possible_mappings, uploader_name, overwrite_in_prog)
 
 
-def import_data(path: Path, data_source: DataSource, uploaded_by: Optional[str] = None, overwrite_in_prog=False):
-    in_progress = import_in_progress(data_source)
+def import_data(path: Path, mappings: List[xlsx_utils.SheetMapping], uploaded_by: Optional[str] = None,
+                overwrite_in_prog=False):
+    error_collector = ErrorCollector()
+    data_file = {mapping.data_file for mapping in mappings}
+    if len(data_file) != 1:
+        raise ImportError("Something is wrong, can't import from two different files...")
+    data_file = mappings[0].data_file
+    in_progress = import_in_progress(data_file)
 
     if in_progress.count() > 0:
         if overwrite_in_prog:
@@ -71,44 +71,32 @@ def import_data(path: Path, data_source: DataSource, uploaded_by: Optional[str] 
         checksum = hashlib.sha256(f.read())
 
     uploaded_by = uploaded_by or ''
-
-    mapping = MAPPINGS[data_source]
-    data = import_xlsx(path, mapping)
-    data = list(data)
-    obj_types = {Purchase, Delivery, Inventory}
-    num_active_objects = {
-        obj_type.__name__: obj_type.objects.prefetch_related('source').filter(
-            source__status=ImportStatus.active, source__data_source=data_source).count()
-        for obj_type in obj_types
-    }
-    import_stats = collections.defaultdict(lambda: 0)
     data_import = DataImport(
         status=ImportStatus.candidate,
-        data_source=data_source,
+        data_file=data_file,
         uploaded_by=uploaded_by,
         file_checksum=checksum,
         file_name=path.name
     )
     data_import.save()
-    for item in data:
-        for obj in item.to_objects():
-            import_stats[obj.__class__.__name__] += 1
-            obj.source = data_import
-            obj.save()
-    print('Before: ', num_active_objects)
-    print('Candidates: ', dict(import_stats))
+
+    for mapping in mappings:
+        data = import_xlsx(path, mapping, error_collector)
+        data = list(data)
+        for item in data:
+            for obj in item.to_objects(error_collector):
+                obj.source = data_import
+                obj.save()
+
+    print(f"Errors: ")
+    print(error_collector)
     return data_import
 
 
 def complete_import(data_import: DataImport):
-    current_active = DataImport.objects.filter(data_source=data_import.data_source, status=ImportStatus.active)
+    current_active = DataImport.objects.filter(data_file=data_import.data_file, status=ImportStatus.active)
     if current_active.count() > 1:
         raise DataImportError('Multiple active imports')
     current_active.update(status=ImportStatus.replaced)
     data_import.status = ImportStatus.active
     data_import.save()
-
-
-def full_import(path: Path, data_source: DataSource, overwrite_in_prog: bool = False):
-    data_import = import_data(path, data_source, overwrite_in_prog=overwrite_in_prog)
-    complete_import(data_import)
